requirements.txt
aioredis==1.3.1
fastapi==0.104.1
fastapi-cache2==0.2.1
pydantic==1.10.13
pydantic_core==2.14.5
uvicorn==0.21.1
python-dotenv==1.0.0
celery==5.3.6
flower==2.0.1
redis==5.0.1
elasticsearch==7.15.2
elasticsearch-async==6.2.0